type fixup
git-svn-id: file:///svn/unbound/trunk@1416 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ daemon/remote.c
@@ -1300,7 +1300,7 @@ daemon_remote_exec(struct worker* worker)
 {
 	/* read the cmd string */
 	uint8_t* msg = NULL;
-	size_t len = 0;
+	uint32_t len = 0;
 	if(!tube_read_msg(worker->cmd, &msg, &len, 0)) {
 		log_err("daemon_remote_exec: tube_read_msg failed");
 		return;
